+++ apps/proxy/payment/src/main/java/com/example/application/port/in/CancelCoffeeOrderUseCase.java
@@ -1,0 +1,8 @@
+package com.example.application.port.in;
+
+import com.example.framework.adapter.in.web.dto.in.CancelOrderIn;
+import com.example.framework.adapter.in.web.dto.out.CancelOrderOut;
+
+public interface CancelCoffeeOrderUseCase {
+    CancelOrderOut cancelOrder(CancelOrderIn cancelOrderIn);
+}
