+++ apps/proxy/payment/src/main/java/com/example/application/port/in/QueryCoffeeOrderUseCase.java
@@ -1,0 +1,7 @@
+package com.example.application.port.in;
+
+import com.example.framework.adapter.in.web.dto.out.RetrieveOut;
+
+public interface QueryCoffeeOrderUseCase {
+    RetrieveOut queryOrder(String orderId);
+}
