+++ apps/proxy/payment/src/main/java/com/example/application/service/CreateCoffeeOrderService.java
@@ -1,30 +1,22 @@
 package com.example.application.service;
 
-import com.example.application.port.in.CoffeeOrderInputPort;
-import com.example.application.port.out.LoadCoffeeOrderOutputPort;
+import com.example.application.port.in.CreateCoffeeOrderUseCase;
 import com.example.application.port.out.PaymentGatewayOutputPort;
 import com.example.application.port.out.SaveCoffeeOrderOutputPort;
 import com.example.domain.entity.Payment;
-import com.example.framework.adapter.in.web.dto.in.CancelOrderIn;
 import com.example.framework.adapter.in.web.dto.in.CreateOrderIn;
-import com.example.framework.adapter.in.web.dto.out.CancelOrderOut;
 import com.example.framework.adapter.in.web.dto.out.CreateOrderOut;
-import com.example.framework.adapter.in.web.dto.out.RetrieveOut;
-import com.example.framework.adapter.out.network.response.GraphQLCancelResponse;
 import com.example.framework.adapter.out.network.response.GraphQLPaymentResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class CoffeeOrderUseCase implements CoffeeOrderInputPort {
+public class CreateCoffeeOrderService implements CreateCoffeeOrderUseCase {
     private final SaveCoffeeOrderOutputPort saveCoffeeOrderOutputPort;
-    private final LoadCoffeeOrderOutputPort loadCoffeeOrderOutputPort;
     private final PaymentGatewayOutputPort paymentGatewayOutputPort;
 
     @Override
-    @Transactional
     public CreateOrderOut createOrder(CreateOrderIn orderIn) {
         Payment payment = createOrderInToPayment(orderIn);
 
@@ -51,31 +43,5 @@ public class CoffeeOrderUseCase implements CoffeeOrderInputPort {
                 orderIn.getCoffeeMenu(),
                 orderIn.getCoffeePrice()
         );
-    }
-
-    @Override
-    @Transactional
-    public CancelOrderOut cancelOrder(CancelOrderIn cancelOrderIn) {
-        Payment payment = loadCoffeeOrderOutputPort.loadPayment(cancelOrderIn.getOrderId());
-        GraphQLCancelResponse.CancelResponseData response =
-                paymentGatewayOutputPort.requestCancel(
-                        payment.getOrderId(),
-                        "MIGRATION_CAFE"
-                );
-
-        if ("SUCCESS".equals(response.getResult())) {
-            payment.updateCancel();
-            saveCoffeeOrderOutputPort.save(payment);
-            return CancelOrderOut.createSuccess();
-        } else {
-            payment.updateFail();
-            return CancelOrderOut.createSuccess();
-        }
-    }
-
-    @Override
-    public RetrieveOut queryOrder(String orderId) {
-        Payment payment = loadCoffeeOrderOutputPort.loadPayment(orderId);
-        return new RetrieveOut(payment.getOrderId(), payment.getStatus().toString());
     }
 }
